initial commit to this repository
the extension was firstly integrated into tulama. but the code is
moved to a separated repository. this has the advantage that the code
from the tulama does not have to be integrated into this code base.
both could be developed further without disturbing the other.

implemented are the features user card and label printing.