refactoring
the MutationObserver call from Label is moved to tualpaka. the
intention is to have a observer which listen to changes of the
loadingBlockerStatusIdentifier. therefore it is now possible that
multiple features add a callback to call when the observer fires.

Label.js has now a function to add the button. This is added to the
window.loadingBlockerEvents array. the elements from this array are
called from callback given to the constructor of the MutationObserver.