Label removeElementsForInstituteLabel
if the library key contains a
institute name then it should not be printed. it was intended to add
every institute to the sublibraries list, but this is not usefull for
such a big list. but it was not possible to do it in another way.

now every institute name has the institute number in front. now it is possible
to know that the library name is a institute name. therefore it is now
possible to check against the existence of a number with 4 digits in
front of the name and remove the library and location key if it is so.