Label changes
restructure the code a little bit. it was little bit difficult to see
how the different codes of the different art of labels should behave.
now every art has a class and it should be easy to extend the
different functionalities.

in addition there was a little bugfix for the description thing. for
the journals it was necessary to move the whole block upwards.
therefore if the location is empty, then it is not displayed.