ShuffleSignature insertIntoExemplarSignature bugfix
it seams that alma does not saves the value set to the input tag
directly. it has to be done in the way that the user would use.

it was possible to set the value "Anderes Schema" directly. It was
visible but after saving the whole record the "Anderes Schema" value
was not more here.