UserManagement usability change contextMenu entry -> button usage
to print the user it was for now possible to do it with a right click
to open the context menu on the page and then click on the print user
entry. this option has been removed. to be consistent with the other
features a new button has been added to the page where it is allowed
to print the user. the button has been added to the left of the save
button.